Benchmark for SpectraInfo::getL2()
Comes out with following performance metrics:

Benchmark                                     Time(ns)    CPU(ns) Iterations
SpectrumInfoFixture/BM_l2_access               8053297    8051753         77                7.10658M items/s

+++ cow_instrument/DetectorInfo.h
@@ -122,7 +122,7 @@ public:
   }
 
   double l1() const { return m_l1; }
-  double size() const { return m_nDetectors; }
+  size_t size() const { return m_nDetectors; }
 
   const InstTree &const_instrumentTree() const { return *m_instrumentTree; }
